+++ CHANGELOG.md
@@ -2,6 +2,30 @@
 
 All notable changes to this project will be documented in this file.
 
+## 0.1.8 (2021-11-08)
+
+#### Changes
+
+- Update homebridge after toggle.
+
+## 0.1.7 (2021-11-08)
+
+#### Changes
+
+- Code cleanup
+
+## 0.1.6 (2021-11-08)
+
+#### Changes
+
+- Fixed state polling
+
+## 0.1.5 (2021-11-08)
+
+#### Changes
+
+- Changed to poll on an interval
+
 ## 0.1.4 (2021-10-19)
 
 #### Changes